+++ src/app/app.module.ts
@@ -6,8 +6,11 @@ import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
 import { ProductModule } from './products/product.module';
 import { LayoutModule } from './layout/layout.module';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
+
+import { httpInterceptorProviders } from './core';
 
 @NgModule({
   declarations: [
@@ -18,9 +21,10 @@ import { AppRoutingModule } from './app-routing.module';
     SharedModule,
     ProductModule,
     LayoutModule,
+    HttpClientModule,
     AppRoutingModule
   ],
-  providers: [],
+  providers: [httpInterceptorProviders],
   bootstrap: [AppComponent]
 })
 export class AppModule { }