+++ src/app/core/index.ts
@@ -3,3 +3,5 @@ export * from './services';
 
 export * from './guards/can-deactivate.guard';
 export * from './interfaces/can-component-deactivate.interface';
+
+export * from './interceptors';