+++ src/app/core/services/auth.service.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
 
+import { AppSettingsService } from './app-settings.service';
+
 import { Observable, of } from 'rxjs';
-import { delay, tap } from 'rxjs/operators';
+import { delay, switchMap, tap } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -10,10 +12,15 @@ export class AuthService {
   isLoggedIn = false;
   redirectUrl: string;
 
-  login(): Observable<boolean> {
+  constructor(
+    private appSettingsService: AppSettingsService
+  ){}
+
+  login(): Observable<any> {
     return of(true).pipe(
       delay(1000),
-      tap(val => this.isLoggedIn = val)
+      tap(val => this.isLoggedIn = val),
+      switchMap(() => this.appSettingsService.uploadConfigInfo()),
     );
   }