+++ src/app/products/services/index.ts
@@ -1,0 +1,2 @@
+export * from './http-product.service';
+export * from './http-product-observable.service';